refactor: split into modules
