+++ src/error.rs
@@ -1,0 +1,54 @@
+/// Errors returned by the wallet API.
+#[derive(Debug, thiserror::Error)]
+pub enum OdysseyWalletError {
+    /// The transaction value is not 0.
+    ///
+    /// The value should be 0 to prevent draining the service.
+    #[error("tx value not zero")]
+    ValueNotZero,
+    /// The from field is set on the transaction.
+    ///
+    /// Requests with the from field are rejected, since it is implied that it will always be the
+    /// service.
+    #[error("tx from field is set")]
+    FromSet,
+    /// The nonce field is set on the transaction.
+    ///
+    /// Requests with the nonce field set are rejected, as this is managed by the service.
+    #[error("tx nonce is set")]
+    NonceSet,
+    /// The to field of the transaction was invalid.
+    ///
+    /// The destination is invalid if:
+    ///
+    /// - There is no bytecode at the destination, or
+    /// - The bytecode is not an EIP-7702 delegation designator
+    #[error("the destination of the transaction is not a delegated account")]
+    IllegalDestination,
+    /// The transaction request was invalid.
+    ///
+    /// This is likely an internal error, as most of the request is built by the service.
+    #[error("invalid tx request")]
+    InvalidTransactionRequest,
+    /// The request was estimated to consume too much gas.
+    ///
+    /// The gas usage by each request is limited to counteract draining the services funds.
+    #[error("request would use too much gas: estimated {estimate}")]
+    GasEstimateTooHigh {
+        /// The amount of gas the request was estimated to consume.
+        estimate: u64,
+    },
+    /// An internal error occurred.
+    #[error(transparent)]
+    InternalError(#[from] eyre::Error),
+}
+
+impl From<OdysseyWalletError> for jsonrpsee::types::error::ErrorObject<'static> {
+    fn from(error: OdysseyWalletError) -> Self {
+        jsonrpsee::types::error::ErrorObject::owned::<()>(
+            jsonrpsee::types::error::INVALID_PARAMS_CODE,
+            error.to_string(),
+            None,
+        )
+    }
+}
